added a centralized backend logic and global state provider for the socket to be able to use on the whole dashboard component
