+++ chathaven/src/pages/dashboard.js
@@ -1,9 +1,12 @@
 import DashboardPage from "../components/Dashboard";
+import { SocketProvider } from "../components/SocketContext";
 
 export default function Dashboard() {
   return (
-    <div>
-      <DashboardPage />
-    </div>
+    <SocketProvider>
+      <div>
+        <DashboardPage />
+      </div>
+    </SocketProvider>
   );
 }
